Experimental try of composer attachment
